backend/app/services/llm_service.py
from langchain_groq import ChatGroq
from langchain_core.prompts import PromptTemplate
from langchain_core.output_parsers import JsonOutputParser
import os
from dotenv import load_dotenv

load_dotenv()

class LLMService:
    def __init__(self):
        # Initialize LLM (Groq)
        # Ensure GROQ_API_KEY is set in .env
        api_key = os.getenv("GROQ_API_KEY")
        if not api_key:
            print("Warning: GROQ_API_KEY not found in environment variables.")
            self.llm = None
        else:
            self.llm = ChatGroq(temperature=0, model_name="llama3-70b-8192", api_key=api_key)

    def categorize_email(self, email_content: str, prompt_template: str):
        if not self.llm:
            return {"error": "LLM not initialized"}
        
        # TODO: Implement categorization logic
        pass

    def extract_action_items(self, email_content: str, prompt_template: str):
        if not self.llm:
            return {"error": "LLM not initialized"}
            
        # TODO: Implement action extraction logic
        pass

    def generate_draft(self, email_content: str, instructions: str, prompt_template: str):
        if not self.llm:
            return {"error": "LLM not initialized"}
            
        # TODO: Implement draft generation logic
        pass
