backend/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-community
chromadb
sqlalchemy
pydantic
python-multipart
python-dotenv
httpx
